Rename ui to views
The commit renames the `ui` directory to `views` and updates the import paths accordingly. It also removes the `home_page` directory and its contents, as well as the `home_page_viewModel.dart` and `map_model.dart` files.

The `flutter_map_opentopo_polyline.dart` widget is updated to initialize the `mapController` only if it's null, and the settings icon is changed from `Icons.settings_outlined` to `Icons.settings`.

The `home_page.dart` widget is updated to use `FlutterMapOpentopoPolyline` without a `FutureBuilder` when running on the web.