fix map init
The `MapController` was being initialized twice, leading to issues.
This commit fixes the initialization and introduces a `widget_ready` flag in the `MapViewModel` to track if the map controller can be used.

Also includes a minor rename of `to_polyline` to `toPpolyline` in `IgcFlight`.
Adds `sqlite3_flutter_libs` as a dependency for caching map tiles.